delete duplicate UserGql.ts + fix import
