+++ front/src/components/AddFriendsCard/index.tsx
@@ -1,11 +1,11 @@
-import { useQuery } from '@apollo/client';
-import { FC, useContext } from 'react';
-import { UserListQuery } from '../../types/user';
-import { GET_ALL_USERS } from '../../gql/UsersGql';
-import { CircularProgress, Typography } from '@mui/material';
-import CarbonCard from '../CarbonCard';
-import UserList from './components/UserList';
-import { LoginContext } from '../../hooks/useLoginContext';
+import { useQuery } from "@apollo/client";
+import { FC, useContext } from "react";
+import { UserListQuery } from "../../types/user";
+import { GET_ALL_USERS } from "../../gql/UserGql";
+import { CircularProgress, Typography } from "@mui/material";
+import CarbonCard from "../CarbonCard";
+import UserList from "./components/UserList";
+import { LoginContext } from "../../hooks/useLoginContext";
 
 const AddFriendsCard: FC = () => {
   const { data, error, loading } = useQuery<UserListQuery>(GET_ALL_USERS);
